Binary file C project
includes several parts.
you are given a binary file films.bin which contains records of films. Each film record will be
stored using filmData struct.

films.bin file consists of 100 records. 95 of them are blank records. The other 5 records (not blank) are
placed in specific positions of the binary file based on their ids. For example, if id of a film is 5, its record
is the fifth record in the file. The size of each record is equal to the size of filmData struct. The
records are sorted according to their id number.